Rewrite `getContextOptions`
